+++ raster.h
@@ -15,12 +15,12 @@ public:
     Raster(const size_t width, const size_t height);
     ~Raster();
     char getPixel(const unsigned int x, const unsigned int y) const;
-    void setPixel(const unsigned int x, const unsigned int y, char p);
+    void setPixel(const unsigned short x, const unsigned short y, char p);
     std::vector<char> getCol(unsigned int x) const;
     std::vector<char> getRow(unsigned int y) const;
     void clear();
 private:
-    unsigned int pixelIndex(const unsigned int x, const unsigned int y) const;
+    unsigned int pixelIndex(const unsigned short x, const unsigned short y) const;
 };
 
 #endif // __RASTER_H__
